server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlmodel==0.0.14
aiosqlite==0.19.0
aiomysql==0.2.0
pymysql==1.1.1
python-dotenv==1.0.0
httpx==0.25.2
pyyaml==6.0.1
regex==2023.10.3
loguru==0.7.2
typing-extensions==4.8.0
greenlet==3.2.4
qdrant-client==1.7.0
numpy==2.3.2
faiss-cpu==1.12.0
